Extract CST builder
